Agent/requirements.txt
google-adk>=0.4.0
python-dotenv>=1.0.0
litellm>=1.69.0
toolbox-core>=0.1.0
fastmcp>=2.3.4
mysql-connector-python>=9.3.0
tabulate>=0.9.0
typing>=3.10.0.0
pyyaml>=6.0
